spaceflights_with_reporting_and_experiment_tracking/kedro-tutorial/src/requirements.txt
black~=22.0
flake8>=3.7.9, <5.0
ipython>=7.31.1, <8.0
isort~=5.0
jupyter~=1.0
jupyterlab_server>=2.11.1, <2.16.0
jupyterlab~=3.0
kedro-datasets[pandas.CSVDataSet, pandas.ExcelDataSet, pandas.ParquetDataSet]~=1.0.0
kedro-datasets[plotly.PlotlyDataSet, plotly.JSONDataSet]~=1.0.0
kedro-datasets[matplotlib.MatplotlibWriter]~=1.0.0
kedro-telemetry~=0.2.0
kedro-viz>=5.2
nbstripout~=0.4
pytest-cov~=3.0
pytest-mock>=1.7.1, <2.0
pytest~=6.2
scikit-learn~=1.0